[I25-171] feat : PR 리뷰 반영
