+++ src/app/project/[project_name]/components/ProjectMainContent.tsx
@@ -10,10 +10,54 @@ import Markdown from 'markdown-to-jsx';
 import type { MarkdownToJSX } from 'markdown-to-jsx';
 import { Body } from '@/app/components/system/text';
 import type { ProjectDetailViewModel } from './types';
-import {
-  isYoutubeUrl,
-  normalizeYoutubeUrl,
-} from '../services/utils/content-utils';
+
+const normalizeYoutubeUrl = (url: string) => {
+  try {
+    const parsed = new URL(url);
+    if (parsed.hostname.includes('youtu.be')) {
+      return `https://www.youtube.com/embed/${parsed.pathname.replace('/', '')}`;
+    }
+
+    if (parsed.hostname.includes('youtube.com')) {
+      if (parsed.pathname === '/watch') {
+        const videoId = parsed.searchParams.get('v');
+        return videoId ? `https://www.youtube.com/embed/${videoId}` : url;
+      }
+
+      if (parsed.pathname.startsWith('/embed/')) {
+        return url;
+      }
+
+      const pathSegments = parsed.pathname.split('/').filter(Boolean);
+
+      if (parsed.pathname.startsWith('/shorts/') || parsed.pathname.startsWith('/live/')) {
+        const videoId = pathSegments[1] ?? pathSegments[0];
+        return videoId ? `https://www.youtube.com/embed/${videoId}` : url;
+      }
+
+      if (parsed.pathname.startsWith('/v/')) {
+        const videoId = pathSegments[1] ?? pathSegments[0];
+        return videoId ? `https://www.youtube.com/embed/${videoId}` : url;
+      }
+    }
+  } catch (error) {
+    console.error('Failed to normalize YouTube url', error);
+    return url;
+  }
+
+  return url;
+};
+
+const isYoutubeUrl = (url: string) => {
+  try {
+    const { hostname } = new URL(url);
+    const normalizedHost = hostname.replace(/^www\./, '');
+
+    return ['youtube.com', 'm.youtube.com', 'youtu.be'].includes(normalizedHost);
+  } catch {
+    return false;
+  }
+};
 
 const iframeWrapperClass =
   'relative mb-[0.875rem] w-full overflow-hidden bg-black';
