+++ src/app/project/[project_name]/components/project-sidebar/ProjectSummarySection.tsx
