+++ src/app/project/[project_name]/components/project-sidebar/styles.ts
